Add files via upload
How can I ensure that my code, program, or software is functional and secure?
Ensuring your code, program, or software is functional, is a vital aspect of coding, as a programmer your job can be trusted which is very important, therefore to ensure all this is acquired, the first rule is to stick to the requirements and make sure that all are met with no exceptions, another important step would be constant testing of the code to make sure there are no errors or room for errors left open. 


How do I interpret user needs and incorporate them into a program?
Interpreting users' needs as simple as it sounds, could be easily messed up, so you need to be careful while doing that, User's needs can be interpreted by first collecting the user's requirements, after all this is done, you then read it and make sure all needed requirements giving are collected, then you lay it down in an orderly manner of approach, and come up with a plan to meet all requirement step by step, this way there is less room for error and making sure all requirements are well interpreted and met. 


How do I approach designing software?